+++ packages/mechanic/app/functions.js
@@ -1,3 +1,3 @@
-import { setUp } from "../src/function-set-up";
+import { setUp } from "../src/function-set-up.js";
 import functions from "./FUNCTIONS";
 setUp(functions);