+++ packages/create-mechanic/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 0.2.10 - 2021-07-21
+
+### Fixed
+
+- Fixed Canvas Image example param definition
+
 ## 0.2.7 - 2021-07-13
 
 First logged release